rename scheduler clock to InternalClock
