+++ kyo-scheduler/js/src/main/scala/kyo/scheduler/InternalClock.scala
@@ -4,10 +4,10 @@ import java.util.concurrent.Executor
 import scala.annotation.nowarn
 
 @nowarn
-final class Clock(executor: Executor = null):
+final class InternalClock(executor: Executor = null):
 
     def currentMillis(): Long = System.currentTimeMillis()
 
     def stop(): Unit = {}
 
-end Clock
+end InternalClock
