+++ kyo-scheduler/shared/src/main/scala/kyo/scheduler/Task.scala
@@ -4,7 +4,7 @@ trait Task(initialRuntime: Int):
 
     @volatile private var state = Math.max(1, initialRuntime) // Math.abs(state) => runtime; state < 0 => preempting
 
-    private[kyo] def doRun(clock: Clock): Task.Result =
+    private[kyo] def doRun(clock: InternalClock): Task.Result =
         val start = clock.currentMillis()
         try run(start, clock)
         finally state = (Math.abs(state) + clock.currentMillis() - start).toInt
@@ -17,7 +17,7 @@ trait Task(initialRuntime: Int):
     final def preempt(): Boolean =
         state < 0
 
-    def run(startMillis: Long, clock: Clock): Task.Result
+    def run(startMillis: Long, clock: InternalClock): Task.Result
 
     def runtime(): Int = Math.abs(state)
 end Task
@@ -41,7 +41,7 @@ private[kyo] object Task:
 
     inline def apply(inline r: => Unit): Task =
         new Task(0):
-            def run(startMillis: Long, clock: Clock) =
+            def run(startMillis: Long, clock: InternalClock) =
                 r
                 Task.Done
             end run
